Add files via upload
It is a Bank Management System written in C . Here I have used files for the creation of an account. This will run on the user's computer to create files which will store all the necessary information regarding that account which the user wants to create. The program titled-"bank" will store the necessary information highlighting the reasons due to which the user had to withdraw , deposit, taking a Loan , etc. It takes a username and a password to access the file. The "Transaction . C" program creates the file where the user will enter the bank name , account details, password, name of the account etc. The "main .C" will access those files and do the necessary operations.